move components folder under src, change the import link and errors fixed
