+++ src/components/searchedProduct.vue
