+++ src/views/ProductDetailForBuyer.vue
@@ -55,7 +55,7 @@
 </template>
 
 <script>
-    import messageContent from '../../components/messageContent.vue';
+    import messageContent from '../components/messageContent.vue';
     export default {
     components: { messageContent },
         data() {
